refactor code: rename and move all negotiator code to one package
